refactor: move masstransit filters into filters directory
